Transpile: Separate SCC-CUDA tests into their own sub-package
